+++ Repositories/EmployeeRepository.cs
@@ -10,7 +10,7 @@ public class EmployeeRepository : IEmployeeRepository
     private readonly AppDbContext _db;
     public EmployeeRepository(AppDbContext db) => _db = db;
 
-    public async Task<Employee?> GetByIdAsync(int id) =>
+    public async Task<Employee?> GetByIdAsync(long id) =>
         await _db.Employees.AsNoTracking().FirstOrDefaultAsync(e => e.Id == id);
 
     public async Task<IReadOnlyList<Employee>> ListAsync(Expression<Func<Employee, bool>>? predicate = null)
